Fully working version
This Version is fully working for the following asset types:
Compute Engine
Compute Engine Disks
CloudSQL
Buckets
Big Query Dataset
