+++ Requirements.txt
@@ -1,1 +1,5 @@
+# Function dependencies, for example:
+# package>=version
 google-api-python-client
+google-auth
+google-cloud-bigquery
